+++ Sources/SentryCrash/Recording/Monitors/SentryCrashMonitor_System.h
@@ -39,6 +39,8 @@ SentryCrashMonitorAPI *sentrycrashcm_system_getAPI(void);
 
 bool sentrycrash_isSimulatorBuild(void);
 
+uint64_t sentrycrashcm_system_freememory(void);
+
 #ifdef __cplusplus
 }
 #endif