+++ Tests/SentryTests/SentryCrash/TestSentryCrashWrapper.m
@@ -14,6 +14,7 @@
     instance.internalIsApplicationInForeground = YES;
     instance.installAsyncHooksCalled = NO;
     instance.closeCalled = NO;
+    instance.internalFreeMemory = 0;
     return instance;
 }
 
@@ -55,6 +56,11 @@
 - (NSDictionary *)systemInfo
 {
     return @{};
+}
+
+- (NSInteger)freeMemory
+{
+    return self.internalFreeMemory;
 }
 
 @end